requirements.txt
streamlit
pandas
numpy
scikit-learn
pydantic
langchain
langchain-openai
openai
python-dotenv